refactored code in home page
